% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{bsaR-class}
\alias{bsaR-class}
\title{bsaR class}
\description{
bsaR class
}
\section{Slots}{

\describe{
\item{\code{meta}}{data.frame}

\item{\code{AD}}{matrix}

\item{\code{Depth}}{matrix}

\item{\code{Freq}}{matrix}

\item{\code{window}}{list}
}}

