% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcMethods.R
\name{CalcFreqByWindow}
\alias{CalcFreqByWindow}
\title{caculate the frequency by windows}
\usage{
CalcFreqByWindow(bsa, widow.size = 50000)
}
\arguments{
\item{bsa}{bsaR object}

\item{window.size}{window size, default is 50000}
}
\description{
caculate the frequency by windows
}
