% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallGenotype.R
\name{CallGenotype}
\alias{CallGenotype}
\title{call geneotyep by allele frequency}
\usage{
CallGenotype(bsa, low.AF = 0.2, high.AF = 0.8)
}
\arguments{
\item{bsa}{bsaR object}

\item{low.AF}{low allele frequency threshold, default is 0.2}

\item{high.AF}{high allele frequency threshold, default is 0.8}
}
\description{
The function will call genotype by allele frequency
0 means homozygous reference, 1 means homozygous alternative, 2 means heterozygous
}
