% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMethods.R
\name{FilterByAF}
\alias{FilterByAF}
\title{filter allele matrix by low AF}
\usage{
FilterByAF(bsa, min.AF = 0.01, max.AF = 1)
}
\arguments{
\item{bsa}{bsa object}

\item{min.AF}{minimum AF threshold, default is 0.01}

\item{max.AF}{maximum AF threshold, default is 0.01}
}
\description{
filter allele matrix by low AF
}
