% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMethods.R
\name{FilterByDepth}
\alias{FilterByDepth}
\title{filter allele matrix by depth}
\usage{
FilterByDepth(bsa, min.depth = 20, max.depth = 3 * min.depth)
}
\arguments{
\item{bsa}{bsa object}

\item{min.depth}{minimum depth threshold, default is 20}

\item{max.depth}{maximum depth threshold, default is 3 * min.depth}
}
\description{
filter allele matrix by depth
}
