% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMethods.R
\name{FilterByGeno}
\alias{FilterByGeno}
\title{filter loci  by geno type}
\usage{
FilterByGeno(bsa, p.wt, p.mut)
}
\arguments{
\item{bsa}{bsa object}

\item{p.wt}{sample name of parent wild type}

\item{p.mut}{sample name of parent mutant}
}
\description{
filter loci  by geno type
}
