% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMethods.R
\name{FilterByPos}
\alias{FilterByPos}
\title{filter allele matrix by Position}
\usage{
FilterByPos(bsa, pos)
}
\arguments{
\item{bsa}{bsa object}

\item{pos}{position vector, e.g. chr1_1000}
}
\description{
filter allele matrix by Position
}
