% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMethods.R
\name{FilterLowDepth}
\alias{FilterLowDepth}
\title{filter allele matrix by depth}
\usage{
FilterLowDepth(bsa, depth = 20)
}
\arguments{
\item{bsa}{bsa object}

\item{depth}{depth threshold, default is 20}
}
\description{
filter allele matrix by depth
}
