% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotWindowFreq}
\alias{plotWindowFreq}
\title{plot window frequency}
\usage{
plotWindowFreq(bsa, chr = NULL, geno1 = NULL, geno2 = NULL,
  window.size = NULL, cols = c("red", "blue", "black"))
}
\arguments{
\item{bsa}{bsa object}

\item{chr}{chromosome}

\item{geno1}{genotype 1}

\item{geno2}{genotype 2}

\item{window.size}{widows.size}
}
\description{
plot window frequency
}
