% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMethods.R
\name{SubsetGenome}
\alias{SubsetGenome}
\title{subset genome}
\usage{
SubsetGenome(bsa, chr, start = -Inf, end = Inf)
}
\arguments{
\item{chr}{chromosome}

\item{start}{start position}

\item{end}{end position}
}
\description{
subset genome
}
