% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcMethods.R
\name{CreateBsaFromVcf}
\alias{CreateBsaFromVcf}
\title{Create bsa object from vcf}
\usage{
CreateBsaFromVcf(vcf.file)
}
\arguments{
\item{vcf.file}{vcf file path}
}
\description{
Create bsa object from vcf
}
